Ruby: consider ':' at the end of word when matching a keyword
Partially close #3306.

The original code wrongly recognize "for" in a hash literal
like { for: nil } as a keyword. When the parser finds "for" keyword,
the parser expects a block comes after the keyword. However, a block
doesn't come after the hash literal. This causes one of the reason
making a loop structure in a scope field of a tag.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>