main: verify the requirement for a memory stream when pushing a new stream
By calling makePromise() with passing (0, 0, 0, 0) as arguments
specifying the area, you can run a guest parser on the whole the area
of the current input file.

Internally, the main part reuses the current input stream for the
guest parser.

I found there is a case that it is impossible to reuse the current
input stream in the guest parser; the current input stream is
opened as as FILE based MIO and the guest parser needs a MEMORY
based MIO.

In this case, the main part should abort reusing and open a new MEMORY
based MIO derived from the underlying FILE based MIO.  However,
reopening is not easy because the main part has no reliable way for
calculating the area information for the MEMORY based MIO.  The
calculation is needed because the caller just passes (0, 0, 0, 0) as
the area information.

Till we can find a good way for calculating the area, what we can do
is just printing warnings.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>