CPreProcessor: run Asm parser on the __ASSEMBLER__ (or __ASSEMBLY__) areas
The issue fixed with this commit is found in

  https://stackoverflow.com/questions/71174162/cant-generate-tag-mrs-s-from-define-mrs-sv-r-in-a-file-even-using/71228230#71228230

The info page of cpp says:

    '__ASSEMBLER__'
	 This macro is defined with value 1 when preprocessing assembly
	 language.

__ASSEMBLY__ is the convention used in Linux kernel source tree.

  https://stackoverflow.com/questions/28924355/gcc-assembler-preprocessor-not-compatible-with-standard-headers

The area may define gas macros that can make C/C++ parser confused.
So this change hides the area from C/C++ parser as if '#if 0' is
specified.  Instead, the CPreProcessor runs Asm parser on the area as
a guest.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>