buildsys: improve docs/Makefile
when user just run `make`, no action name is given, we should show help messages.

checking if sphinx-build is installed is involved in certain actions, not all,
for example, if user run `make clean` or `make help`, we do not need to check sphinx-build if is installed.

old code use `which` command, `which` command is not a builtin command in some shell. it is not pre-installed in some os.
we usually use `command` command to check if a command is exist.

I removed many lines `@echo`, becuase per line shell code runs in a seperate shell.
I removed them let them run in one shell, it can improve performance.

TODO: report this to sphinx project.

Signed-off-by: leleliu008 <leleliu008@gmail.com>