Cxx: use a trashbox for deleteing tokens allocated for anonymous parameters
Partially close #3372.

The original code assumed tokens allocated in
cxxParserTokenChainLooksLikeFunctionParameterList() for tagging
anonymous parameters were deleted on its caller side
(cxxParserEmitFunctionParameterTags()).

However, in some conditions, cxxParserEmitFunctionParameterTags() are
not called. As a result, the allocated tokens are not deleted.

To avoid the memory leaking, the deletion cannot depend on
cxxParserEmitFunctionParameterTags(). This change uses per-parsing
trashbox for deleting the anonymous tokens. The objects linked to
the trashbox are destroyed at the end of parsing the current input
file.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>