Delete tokens allocated for anonymous parameters extracted from a broken parameter list
Close #3372.

The original code assumed tokens allocated in
cxxParserTokenChainLooksLikeFunctionParameterList() for tagging
anonymous parameters are deleted on its caller side.  This assumption
is true if the parameter list where the anonymous parameters are
defined is valid.

If the parameter list is broken, the tokens are never deleted. This
cause a memory leak, as reported in #3372.

With this change, cxxParserTokenChainLooksLikeFunctionParameterList()
releases the tokens if given parameter list is broken.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>