C++: propagate bGotTemplate to lower parser functions
Close #3387.

g_cxx.pTemplateTokenChain represents whether the current parsing
context is in a template or not. However, g_cxx.pTemplateTokenChain
is reset in cxxParserNewStatement(). As the result, the C++ parser
cannot parse template arguments passed with <> in a template well.

    template<typename T>
    class C
    {
      C<T>(void) {
      }
    };

In this input, the parser could not extract C in "C<T>".

This change introduce parameters representing whether the current
parameter context is in a template or not to some parser functions.
So even if g_cxx.pTemplateTokenChain is reset with
cxxParserNewStatement(), the parser can extract C in "C<T>".

Signed-off-by: Masatake YAMATO <yamato@redhat.com>