Cxx: add missing NULL check when detecting a constructor
Close #3417.

The original issue was found when parsing cuda code.
However, the bug is common between cuda and C++.

When detecting a name defined in a class is a constructor
or not, strrstr() is used. The original code assumed the
value returned from strrstr is not NULL. This wrong assumption
caused a crash.

@ijpq made efforts to find an input file causing the crash.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>