javascript: fix the definition of a class
Before ES6, there was no way to tell if an individual function
was a class constructor.  This is why the convention was set that
class names start with uppercase, and function names start with
lowercase.

This patch also stops tagging objects as classes.  Objects in
JavaScript are just like any other value.  One cannot "new" an
object, only a constructor function.