GitHub Actions: serialize the testing on BSDs
We have gotten "429 Too Many Requests" failures when running test cases
on the BSD instances.

While monitoring which one fails, I found an ambiguous rule. We have 4
BSD targets. Only one or two of them got the 429.

My guess of the reason for 429 is that multiple BSD targets start at
the same time or within a short period; the volume of the download
access from vagrant cloud reaches the limit formulated by Vaygrant
Cloud.

This experimental change serializes the execution of tests on BSDs
to limit the volume.

@leleliu008 provides the actual way for serializing at
https://github.com/universal-ctags/ctags/pull/3439#issuecomment-1183709443