Rake: accept symbols defined with single and double quote characters
Close #3474.

A symbol defined with single and double quote characters like :'name'
and :"name" caused a crash of the Rake prser.

The original code handled the initial character, <:> of a name like
<:name> in the Rake parser side. The Rake passed the rest of string
(<name>) to rubyParseMethodName() exported from the Ruby parser.

However, rubyParseMethodName() itself has an ability to handle the initial
character. So this commit makes the Rake parser pass whole <:name> to
rubyParseMethodName(). The single and double quote characters after <:>
is handled in rubyParseMethodName().

Signed-off-by: Masatake YAMATO <yamato@redhat.com>