main: extend the bucket array of hashTable dynamically
There is a plan to store much more items in a hashtable (htable).

The current implementation, allocating a bucket array of a htable with
the size specified when initializing the htable, doesn't fit the
purpose because proper size is not predictable.

With this change, the htable itself extends the size of the
array twice dynamically when the approximate filling rate of the bucket
array of the htable is greater than 80%.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>