main,Zsh: detect zsh in the parser guessing based on emacs modelines (the first line)
We introduced Zsh parser. To utilize the new paresr, it is not enough
to detect a mode in an emacs modeline.

For an example:

   -*- mode: sh; eval: (sh-set-shell "zsh") -*-
   ...

The original code tells that ctags should use "sh" parser for
processing the input.

With this change, ctags can detect "zsh" passed as an argument to
sh-set-shell function.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>