libutil: revise functions for updating the value for a given key
The original hashTableUpdateItem() replaces not only the value
but the key. So we cannot add "const" modifier to the key parameter.

This change revises the function.

The new hashTableUpdateItem() never replaces the key; if key is
not found, it just returns false. So we can add the const modifier
to the key parameter.

In addition we introduces a new function, hashTableUpdateOrPutItem().
hashTableUpdateOrPutItem() don't replace the key as far as the key
exits in the hash table. However, we don't add the const modifier
to the key parameter because the key and the value is put to the
the hash table if the key doesn't exist in the hash table.

These two are tested in tutil target.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>