Fortran,refactor: add new global variable representing parsing pass
currentPass is the new variable representing parsing pass.
Instead, FreeSourceFrom, a global variable is removed.

FreeSourceFrom had two aspects: the pass of parsing (1 or 2) and
the expected input syntax (free or fixed). It makes understanding the
source hard.

currentPass focuses on the pass only.
Additional introduced macros, inFreeSourceForm and inFixedSourceForm
referring to currentPass internally are for representing
the expected input syntax.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>