json-writer: provide the way to define extra key and value pair
With the original code, json-writer emits the following form for
TAG_ROLE_DESCRIPTION:

      {"_type": "ptag", "name": "TAG_ROLE_DESCRIPTION",
			"parserName": "LANG!KIND", ...

This commit is PREPARATION for changing the form like:

      {"_type": "ptag", "name": "TAG_ROLE_DESCRIPTION",
			"parserName": "LANG",
			"kindName": "KIND", ...

In the "parserName" field, two values are embedded.  With this change,
by specifying jsonObjectKey in the ptag definition, json-writer split
the value for parserName into two subvalues.  The first subvalue is
for "parserName" key, and the second subvalue is for the key specified
with jsonObjectKey.

This extension is for newer file format. So it is disabled now.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>