writer-json: don't escape / with backslash in pattern fields of ptags
/ in pattern fields was converted to \/ in json format.
This conversion comes from the format of tags(5). In the tags format,
/ in pattern fields is used as a separater of the field like:

  /int main(/

So a / in the middle of the field must be converted to \/.

However, in the json format doesn't use such a separater. The
conversion is not needed.

NOTE: in a regular tag, the conversion is still needed even in the json
format because / and ? are used as not only separaters but also indicators
of searching direction like:

  /int main(/
  ?int main(?

Signed-off-by: Masatake YAMATO <yamato@redhat.com>