sql: parse PL/SQL selection directives
PLSQL selection directives are of the form

$IF boolean_static_expression $THEN
   text
[ $ELSIF boolean_static_expression $THEN
   text
]...
[ $ELSE
   text
]
$END

In addition, boolean_static_expression conditions like

$$MY_CUSTOM_VAR > 3

can appear which confuses the current parser because it
thinks $$ is the start of dollar-quoted string and drops
the rest of the code.

This patch keeps only the $IF branch and skips all the code
between $ELSE or $ELSIF and $END. In addition, this patch
also drops $ from all identifiers starting with $$ behind
$IF and $ELSIF as these may be conditional variables and
not dollar-quoted strings.