main: Don't strdup the inputFileName when storing a tag to the corkQueue
When storing a tag entry to the corkQueue, the most of members
of the tag entry are duplicated for isolating the the stored tag
entry from the original parsing context.

In the most of all cases, we can expect all tag entries in the
corkQueue has the same value as inputFileName member. Strdup'ing
the value is waste of memory resource especially when the input
file name is long.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>