vstring: Avoid int -> char truncation warnings
Either avoid or silence expected casts from int to char.

I'm not entirely sure whether vStringPut() shouldn't actually take a
char argument instead of an int, as it will cast it in all cases. It is
however slightly more convenient as the APIs to retrieve characters
uses int only for edge cases (e.g. EOF), so the common case is a safe
cast and it's more convenient not to explicit it everywhere.