main: put tags having "end:" field to interval tree
We call the interval tree "intervaltab".

If the cork queue is enabled in a parser, a tag entry
having non-zero value as its "end:" field is stored to
the intervaltab.

We can search tags in intervaltab with

   int queryIntervalTabByLine(unsigned long lineNum),
   int queryIntervalTabByRange(unsigned long start, unsigned long end), or
   int queryIntervalTabByCork(int corkIndex).

Signed-off-by: Masatake YAMATO <yamato@redhat.com>
(cherry picked from commit 4a2d580f72f02065134400eb7a0d1d560231c609)