main: delay reading all inputs when the parser just prepares its base parser
This commit fixes a bug appeared in SUBPARSER_SUB_RUNS_BASE
base/subparser combination where --regex-<LANG> is used for
implementing the subparser.

In such a combination, the subparser may call
scheduleRunningBaseparser() in its "parser" method.
Calling scheduleRunningBaseparser is only for running
the base parser. The base parser may run the subparser.

The expected steps are <E>:
1. A subparser schedules running its base parser.
2. The base parser runs.
   The patterns of --lang-<SUBPARSER> options are applied during in
   this step.

However, the original code did <W>:
1. A subparser schedules running its base parser.
   The patterns of --lang-<SUBPARSER> options are applied after
   scheduling.
2. The base parser runs.

This out of order execution didn't issue because the extracted tags
are the same in <E> and <W>. However, the execution can be a trouble
if the application of the pattern of --lang-<SUBPARSER> option
accesses the information of tags already extracted by the base parser.

In <E>, there are no information of tags when applying the
patterns because the base parser is not run yet at the step 1.

Currently ctags doesn't provide the way for accessing the information
of tags already extracted by a base parser to its subparser. But we
have a plan to introduce such a feature in soon.

TODO: a test case

Signed-off-by: Masatake YAMATO <yamato@redhat.com>
(cherry picked from commit efe6413e0f24110e27b7e97ec09d5ed16f2f2ac0)