main: consider foreign languages when solving the field type specifying _field regex flag
The original code assumed the parser specified in a regex pattern
was the owner of the field specified with the _field regex flag in
the pattern.

For an example, in the option

    --regex-Foo=/.../{_field=afield:\1}...

, ctags assumed "afield" was a specific field of "Foo" parser.
The original code didn't consider an option like

    --regex-Foo=/.../{_language=Bar}{_field=afield:\1}...

In this case, ctags should consider "afield" as a specific
field of "Bar" parser.

NOTE: the language specified with _language flag must be
declared with _foreignLanguage= in --langdef-Foo.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>