main: define operators accessing parser specific fields
The operators (setter and/or getter) for each field.
The name of operatos has following form:

	setter => :<LANG>.<FIELD>
	getter => <LANG>.<FIELD>:

NOTE: The operators for a field are defined only if the
field has C level getters and setters, getValueObject and setValueObject
in fieldDefinition. This commit doesn't provide a getter and a setter.
The test cases and the real accessor will be implemented later.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>