main/field: define operators accessing parser specific fields
The name of operators has following form:

	setter => <LANG>.<FIELD>:
	getter => :<LANG>.<FIELD>

The operators for a field are automatically defined only
if the field has a setter and a getter in C language level;
the operators require that the fieldDefinition for the field
has getValueObject method setValueObject method.

Currently, no parser has such methods.  So, we cannot test the new
code added with this commit yet.

In later commits, I will implement the methods for fields defined with
--_fielddef option.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>