main: count the added tags after emitting parser-specific ptags
Close #3935.

C++ may return RESCAN_FAILED when giving a broken input.
Just "private" is an example of a broken input:

	echo private > input.cpp

When a parser returns RESCAN_FAILED, the main part must reset the
state of parsing to the initial state. The main part records the
initial state for the case when the reset is needed.

The bug comes from the code not implemented as the expectation:

EXPECTATION:
* The state includes the position of tags output stream AND the number
  of emitted tags.

* The "initial state" should be recorded at the time before running
  the parser and after emitted the parser-specific ptags.

ORIGINAL CODE: The number of emitted tags is recorded before the main part
emitting the parser-specific ptags.

internalSortTags() takes the number of emitted tags. With the original,
the number can become incorrect when the a parser returns RESCAN_FAILED.

Acknowledgments:
@terminatorul found the minimum input for reproducing the bug.
@leleliu008 found --disable-external-sort as the critical condition
for reproducing the bug.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>