main: consider foreign languages when solving the extra specifying _extra regex flag
The original code assumed the parser specified in a regex pattern
was the owner of the extra specified with the _extra regex flag in
the pattern.

For an example, in the option

    --regex-Foo=/.../{_extra=anextra}...

, ctags assumed "anextra" was a specific extra of "Foo" parser.
The original code didn't consider an option like

    --regex-Foo=/.../{_language=Bar}{_extra=anextra}...

In this case, ctags should consider "anextra" as a specific
extra of "Bar" parser.

NOTE: the language specified with _language flag must be
declared with _foreignLanguage= in --langdef-Foo.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>