lex: make reference tags for the conditions used for making groups
In some lexers, all the rules for a condition are grouped together as:
<COND>{
...
}

It is useful to be able to jump to these blocks.

Conditions used for grouping are tagged as reference tags having
"grouping" role of "cond" kind. So you must add --extras=+r
to your ctags cmdline for enabling this support.

Update the parser version to 1.1.

This is derrived from #3937.