Asm: relax the condition for accepting characters within a section name
If a section name is built with a macro expansion, the following
strings may appear in parts of the string:

* \param,
* \(), and/or
* \@.

The original code didn't consider the characters in the strings,
so it cannot extract section names from a macro definition.

Though it is debatable to extract a string including references to
macro parameters, ctags may emit the following warning message if we
consider the characters as parts of a section name:

    ctags: Notice: ignoring null tag ... (line: 13, language: LdScript)

By accepting the characters, we can suppress the warning messages.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>