main: don't allocate a buffer for tagEntryInfoX::sourceFileName if possible
When queuing a tagEntryInfo to the cork queue, sourceFileName
was strdup'ed. However, in many cases, the strings for entries
in the queue have the same value.  We can reduce memory allocations
by using the same buffer as we did in the inputFileName member.

Under massif, I measured the peak memory usage of ctags processes
indexing the source tree of the Linux kernel.

Without this change, the peak is 132.7 MiB.
With this change, the peak is 121.9 MiB.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>