build-sys: utilize pegof
Pegof is a PEG grammar optimizer and formatter developed at
https://github.com/dolik-rce/pegof.

This change is for utilizing pegof for optimizing our PEG-based
parsers.

We assume pegof is installed at ${somewhere}/pegof.

  $ ./autogen.sh
  $ ./configure ... --with-pegof=${somewhere}/pegof ...
  $ make

To verify the command line invoking pegof, pass V=1 option to make
like:

  $ make V=1

Makefile runs pegof to generate a .pego file from the given .peg
file. Then, it runs packcc to generate .c and .h files from the
.pego file. Via make command, you can generate a specified pego file
directly. e.g.

  $ make V=1 peg/varlink.pego

This may be helpful for debugging pegof.

If you successfully build ctags with pegof, you will see "pegof" in
the output of ./ctags --list-features.

    % ./ctags --list-features
    #NAME             DESCRIPTION
    ...
    pegof             makes peg based parser(s) optimized (experimental)
    ...

For the debugging purpose, .pego files are in EXTRA_DIST.
In the future, we should remove them. So we don't have to add
.pego files to our .tar.gz, the result of make dist.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>