main: extend --list-languages option to list only parsers using packcc
NOTE: This is a hidden extension. We dont't update the help messages and
the description of the man pages for this extension.

This extension is for choosing parsers for testing.
e.g. running test cases related to PEG-based parsers only:

  make units LANGUAGES="$(./ctags --list-languages=_packcc |  tr '\n' ',')"

As the alternative to "_packcc", you can specify "_libxml" or "_libyaml" as
the argument for the option.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>