JavaScript: generate tags for empty properties with setting allowNullTag
In javascript it is legal to have e.g. {"": false}.  With the
original code the parser attempts to generate the tag for the empty
string which produces warning.

With setting allowNullTag member of tagEntryInfo, the parser can
generate such tags without producing warnings.

Actual tag emission is controlled by --extras=+{nulltag}.

This commit is based on the commit in #4153.

Co-Author: Masatake YAMATO <yamato@redhat.com>