Tmain: add more cases for testing boolean type fields
NOTE: This test case is broken. In xref output, the boolean fields
should be printed as '-' if the field is not set. In tags output, the
value for boolean field should not be printed; if the field has an C
string, the field name should be printed.

The current implementation doesn't support parser specific fields
typed boolean well. The following commits fix this brokenness.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>