CPreProcessor: revise the interface of the macro expander
The original code provided a vString as the result of macro
expansion. In the proces of expansion, we lost the orignal line
numbers of the macro arguments.

To solve "CPreprocessor,C: macro expands multiple lines" #4018, we
must provide a way to keep the the orignal line numbers in the result
of expansion. This commit is the first step for providing the way.

TODO:
- limiting the size of the unget buffer

Signed-off-by: Masatake YAMATO <yamato@redhat.com>