Asm: use cppGetInputLineNumber() and cppGetInputFilePosition()
Just using the fuctions. The improvement with this change is limited.

Unlike Cxx and LdScript parsers, Asm parser is not token oriented.
As a result, we cannot preserve the line number of the
arguments passed to a Cpp macro in some cases.
See th .b test case of this commit.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>