CPreProcessor,LdScript,Asm: (bugfix) don't expand macros defined with -I
Fixed #4225

tags.sh in Linux passes "-I const" to ctags. This caused a crash
of ctags when processing arch/arm64/crypto/aes-neon.S.

-I option defines a macro having no replacement.  cppExpandMacro()
returns NULL for such a macro.  It causes a crash because
cppUngetMacroTokens doesn't consider NULL as an argument.

If a macro has no replacement, we should not call cppExpandMacro()
with the macro. We can avoid the crash and improve the performance a
bit.

Bisected-by: @finallyok
Signed-off-by: Masatake YAMATO <yamato@redhat.com>