lxpath,yaml: (bug fix) adjust line numbers for tags if the parser runs as a guest
Fix #4181.

This commit consists of two interdependent parts that could not be
separated into two individual commits. Introducing either part first
would surface a previously hidden crash—the first sequence triggers
one crash, while introducing the other part first triggers a different
one. Merging any crash-inducing commit would hamper future git bisect
debugging. Therefore, both parts are combined here into a single
commit.

The first part changes main/promise.c and main/read.c.

When a parser made a promise for the same area (thin area) of the
parser, the calculation of the area was delayed till the promise is
forced.

This delaying can be a bug if a guest parser makes a thin area because
the guest parser expects the thin area is the same as the current area
where the guest parser processes.  On the other hand, forcing the
promise is always done in the base area.

With this change, makePromise() does the calculation of the thin area
if the parser runs as a guest.

The second part changes main/lxpath.c and parsers/yaml.c.

The meta parsers user external libraries (libxml2 and libyaml) for
parsing. The libraries provides the way to get the current line
numbers. However, the coordinate system of the line number is the
current coordinate system. updateTagLine() and setTagEndLine() expect
a line number in the absolute coordinate system. In the 2nd part, we
utilize translateLineNumber() for convert the line numbers to the
values in the absolute coordinate system.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>