peg: (bug fix) adjust file offset fot tags if a peg-based parser runs as a guest
Fix #4114.

setTagEndLine() takes a tag entry and a line number as arguments.
setTagEndLine() expects the line number in the absolute coordinate
system. However, a peg-based parser running as a guest parser passed a
line number in the current coordinate system. Using the wrong
coordinate system caused trouble; the end line number was smaller
than the start line number assigned to the tag.

The line numbers were derived from file offsets provided by
the code generated by packcc. The file offsets were in the current
coordinate system. As a result, the derived line numbers were in
the current coordinate system.

This change uses translateFileOffset() to convert the coordinate
system of the offset.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>