TOML,Cargo: disable the parsers temporarily
See #4096.

The TOML parser is broken. I replaced '|' in ABNF with '/' in PEG without
enough consideration when importing the ABNF definition
from https://github.com/toml-lang/toml/blob/main/toml.abnf.

I kept the code because many test cases depend on the TOML parser.
In addition, the design of the kind definition of the parser is
still valuable

The Cargo subparser is also disabled.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>