main: use regular expression for choosing parser
This change extends --map-<LANG> option to support regular
expression matching with the full file name.

The original --map-<LANG> option supports the glob based matching
and the extension comparison with the file basename.
However, two methods are not enough if the file names are too
generic. See https://github.com/universal-ctags/ctags/pull/3287 .

The regular expression passed to --map-<LANG> must be surrounded
by % character like

   --map-RpmMacros='%(.*/)?macros\.d/macros\.([^/]+)$%'

If you want to match in a case-insensitive way, append `i' after
the second % like

   --map-RpmMacros='%(.*/)?macros\.d/macros\.([^/]+)$%i'

If you want to use % as part of an expression, put \ before %
for escaping.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>