mio: update position when using mio_seek() even when pointing behind EOF
When a position passed to mio_seek() is behind EOF, set the internal
pointer to the position corresponding to EOF.

This is how normal fseek() behaves on my machine (probably because it
first has to read the necessary amount of bytes from the stream to get
to the right position). This could help in situations when the return value
of mio_seek() isn't checked and EOF position corresponds "more closely" to
the (invalid) position passed as the parameter instead of staying at
the same position.

Note that when using negative offset and the result is before the
beginning of the file, fseek() doesn't move its internal position
and stays at the same position (probably because it already knows how
many bytes have been read). So this case hasn't been updated in this
patch (though it might be worth considering).