Ada: don't consider "if ~ end if;" and "case ~ end case;" as blocks
The parser calls adaParser() when entering a block.
adaParser() may return when it finds "end ...;" on input
stream.

Though the parser didn't call adaParse() when it found
"if" or "case", adaParse() returned the control when it
found "end if;" or "end case;". This unbalance caused
bugs as reported in #4295.

This change makes adaParse not return the control even
when it finds "end if;" or "end case;".

722d25d introduced the bug fixed in this chhange.

May fix #4295

Signed-off-by: Masatake YAMATO <yamato@redhat.com>