Cxx: don't expand macro if expected macro arguments are not found
Fixes openssl/openssl#28332.

The original code expands `data' in `int data;' in the input.c:

  #define data(A) something
  int data;

So the input becomes:

  int something;

This is a bug in our macro expander. Though the macro definition
expects an argument, no argument list follows `data' in `int data;'.
In such cases, the parser should not expand `data'.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>