Python: parse type statements
Currently this recognizes and skips 'type' at the beginning of a
statement, and then proceeds to parse the rest as a simple variable
assignment.

This works fine for simple types like

    type MyList = list[str]

Inteded to fix #4351.

It doesn't work with parameters like

    type MyList[T] = list[T]

but this is low priority for me, as I don't have any code using those
(yet?), and I'm not used to working in C.